add tests, move tests
