+++ packages/tldraw/src/test/text.test.ts
@@ -1,4 +1,4 @@
-import { TLTextShape, createShapeId } from '@tldraw/editor'
+import { createShapeId } from '@tldraw/editor'
 import { TestEditor } from './TestEditor'
 
 let editor: TestEditor
@@ -265,80 +265,5 @@ describe('When changing text size', () => {
 			x,
 			y,
 		})
-	})
-})
-
-describe('resizing text with autosize true', () => {
-	it('resizes text from the right side', () => {
-		editor.createShape<TLTextShape>({
-			type: 'text',
-			x: 0,
-			y: 0,
-			props: {
-				text: 'Hello',
-				autoSize: false,
-				w: 200,
-			},
-		})
-
-		const shape = editor.getLastCreatedShape()
-
-		const bounds = editor.getShapePageBounds(shape.id)!
-		editor
-			.select(shape)
-			.pointerDown(bounds.maxX, bounds.midY, { target: 'selection', handle: 'right' }) // right edge
-			.expectToBeIn('select.pointing_resize_handle')
-			.pointerMove(bounds.maxX + 100, bounds.midY)
-			.expectToBeIn('select.resizing')
-			.expectShapeToMatch({ ...shape, x: 0, y: 0, props: { w: 300 } })
-	})
-
-	it('resizes text from the right side when alt key is pressed', () => {
-		editor.createShape<TLTextShape>({
-			type: 'text',
-			x: 0,
-			y: 0,
-			props: {
-				text: 'Hello',
-				autoSize: false,
-				w: 200,
-			},
-		})
-
-		const shape = editor.getLastCreatedShape()
-
-		const bounds = editor.getShapePageBounds(shape.id)!
-		editor
-			.select(shape)
-			.keyDown('Alt')
-			.pointerDown(bounds.maxX, bounds.midY, { target: 'selection', handle: 'right' }) // right edge
-			.expectToBeIn('select.pointing_resize_handle')
-			.pointerMove(bounds.maxX + 100, bounds.midY)
-			.expectToBeIn('select.resizing')
-			.expectShapeToMatch({ ...shape, x: -100, y: 0, props: { w: 400 } })
-	})
-
-	it('resizes text from the left side', () => {
-		editor.createShape<TLTextShape>({
-			type: 'text',
-			x: 0,
-			y: 0,
-			props: {
-				text: 'Hello',
-				autoSize: false,
-				w: 200,
-			},
-		})
-
-		const shape = editor.getLastCreatedShape()
-
-		const bounds = editor.getShapePageBounds(shape.id)!
-		editor
-			.select(shape)
-			.pointerDown(bounds.minX, bounds.midY, { target: 'selection', handle: 'left' }) // right edge
-			.expectToBeIn('select.pointing_resize_handle')
-			.pointerMove(bounds.minX - 100, bounds.midY)
-			.expectToBeIn('select.resizing')
-			.expectShapeToMatch({ ...shape, x: -100, y: 0, props: { w: 300 } })
 	})
 })
